chore: replace express with serverless endpoint
