+++ .env.example
@@ -1,3 +1,7 @@
-REACT_APP_BACKEND_URL=
-REACT_APP_ARCHILOGIC_API_URL=https://api.archilogic.com
+ARCHILOGIC_API_URL=https://api.archilogic.com
+ARCHILOGIC_SECRET_KEY=
+BACKEND_URL=http://localhost:3000
+
+REACT_APP_ARCHILOGIC_API_URL=$ARCHILOGIC_API_URL
 REACT_APP_PUBLISHABLE_TOKEN=
+REACT_APP_BACKEND_URL=$BACKEND_URL
