+++ package.json
@@ -20,11 +20,17 @@
     "react-scripts": "3.4.0",
     "redux": "^4.0.5",
     "redux-devtools-extension": "^2.13.8",
-    "redux-thunk": "^2.3.0",  
+    "redux-thunk": "^2.3.0",
     "typescript": "^4.6.4",
-    "uuid": "^7.0.2"
+    "uuid": "^7.0.2",
+    "cors": "^2.8.5",
+    "dotenv": "^8.2.0",
+    "express": "^4.17.1",
+    "http-proxy-middleware": "^1.0.3",
+    "node-fetch": "^2.6.7"
   },
   "scripts": {
+    "vercel": "npx vercel dev",
     "start": "react-scripts start",
     "build": "react-scripts build",
     "test": "react-scripts test",
