--- server/.env.example
@@ -1,2 +1,0 @@
-ARCHILOGIC_SECRET_KEY=
-ARCHILOGIC_API_URL=https://api.archilogic.com
