--- server/.gitignore
@@ -1,11 +1,0 @@
-# dependencies
-/node_modules
-
-.env
-
-# production
-/build
-
-# misc
-.DS_Store
-.env.local
