--- server/package.json
@@ -1,18 +1,0 @@
-{
-  "name": "archilogic-backend",
-  "version": "1.0.0",
-  "description": "Simple server to negotiate a temporary token",
-  "main": "server.js",
-  "scripts": {
-    "test": "echo \"Error: no test specified\" && exit 1"
-  },
-  "author": "mdaguerre",
-  "license": "ISC",
-  "dependencies": {
-    "cors": "^2.8.5",
-    "dotenv": "^8.2.0",
-    "express": "^4.17.1",
-    "http-proxy-middleware": "^1.0.3",
-    "node-fetch": "^2.6.7"
-  }
-}
